#21 scrapes about 56k groceries from wegmans. Tracks state in urls tables, and updated dockerfiles for wegmans spider
