+++ scrapy/Dockerfile.selenium.runtime
@@ -1,0 +1,6 @@
+FROM python
+RUN pip install --upgrade pip; pip install scrapy ; pip install scrapy-selenium ; pip install mysqlclient  ; pip install selenium 
+RUN pip install scrapy-splash # for now we need this in order for it to use the same settings.py
+RUN apt update ; apt install -y firefox-esr  
+ADD https://github.com/mozilla/geckodriver/releases/download/v0.26.0/geckodriver-v0.26.0-linux64.tar.gz /usr/local/bin 
+RUN tar -C /usr/local/bin -xvf /usr/local/bin/geckodriver-v0.26.0-linux64.tar.gz ; rm /usr/local/bin/geckodriver-v0.26.0-linux64.tar.gz ; chmod u+x /usr/local/bin/geckodriver
